rename model_runner.py to model_manager.py
